Reduce runtime hotspot pressure before the next delivery wave
This refactor finishes the current ACP, turn-coordinator, and tools
module decomposition so the post-AgentRuntime runtime spine stops
accumulating more logic inside a few oversized entry files.

It extracts ACP manager/acpx support code, splits provider-turn and
safe-lane paths out of turn_coordinator, moves tool lease/search logic
out of tools/mod, and restores the trusted tool.search visibility and
shell discoverability contracts that broke during the extraction.

The tracked April architecture drift report is refreshed in the same
change so governance reflects the new hotspot measurements instead of
reporting stale pressure.

Constraint: The refactor had to preserve existing runtime/search contracts while shrinking hotspot files
Constraint: Governance requires the tracked monthly architecture drift report to stay fresh after hotspot measurements change
Rejected: Leave tool.search extraction as-is | full all-features verification failed on trusted visibility and shell discoverability regressions
Rejected: Keep the ACP/coordinator/tools logic concentrated in the pre-split files | runtime hotspot pressure would continue to climb and block follow-up work
Confidence: high
Scope-risk: broad
Reversibility: clean
Directive: Preserve trusted _loongclaw tool-search visibility narrowing and shell.exec discoverability semantics when extracting more tool-runtime code
Tested: cargo fmt --all -- --check; CARGO_TARGET_DIR=/tmp/loongclaw-refactor-verify-WqMlDA cargo clippy --workspace --all-targets --all-features -- -D warnings; CARGO_TARGET_DIR=/tmp/loongclaw-refactor-verify-WqMlDA cargo test --workspace --all-features -- --test-threads=1; ./scripts/check_architecture_boundaries.sh; ./scripts/check_dep_graph.sh; ./scripts/check_architecture_drift_freshness.sh docs/releases/architecture-drift-2026-04.md; git diff --check
Not-tested: remote CI
