+++ crates/app/src/acp/acpx.rs
@@ -1,18 +1,16 @@
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeMap;
 use std::io::ErrorKind;
 use std::path::{Path, PathBuf};
 use std::process::Stdio;
 
 use async_trait::async_trait;
-use base64::Engine;
 use serde::{Deserialize, Serialize};
-use serde_json::{Value, json};
-use tokio::io::{AsyncBufReadExt, AsyncReadExt, AsyncWriteExt, BufReader};
 use tokio::process::Command;
-use tokio::time::{Duration, Instant, sleep_until, timeout};
+use tokio::time::Duration;
 
 use crate::CliResult;
 use crate::config::{AcpxMcpServerConfig, LoongClawConfig};
+#[cfg(test)]
 use crate::process_launch::retry_executable_file_busy_async;
 #[cfg(test)]
 use crate::process_launch::retry_executable_file_busy_blocking as retry_spawn_blocking;
@@ -42,8 +40,20 @@ const ACPX_SPAWN_RETRY_ATTEMPTS: usize = 5;
 const ACPX_SPAWN_RETRY_DELAY: Duration = Duration::from_millis(25);
 
 mod command_probe;
+#[path = "acpx_command.rs"]
+mod command_support;
+#[path = "acpx_events.rs"]
+mod event_support;
+#[path = "acpx_handle.rs"]
+mod handle_support;
+#[path = "acpx_process.rs"]
+mod process_support;
 
 use command_probe::{CommandOutputError, wait_for_command_output};
+use command_support::*;
+use event_support::*;
+use handle_support::*;
+use process_support::*;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 struct AcpxRuntimeHandleState {
@@ -70,20 +80,6 @@ struct ResolvedAcpxProfile {
     timeout_seconds: Option<f64>,
     queue_owner_ttl_seconds: f64,
     mcp_servers: BTreeMap<String, AcpxMcpServerConfig>,
-}
-
-#[derive(Debug, Clone)]
-struct AcpxCommandOutput {
-    stdout: String,
-    stderr: String,
-    exit_code: Option<i32>,
-}
-
-#[derive(Debug, Clone, Default)]
-struct AcpxIdentifiers {
-    acpx_record_id: Option<String>,
-    backend_session_id: Option<String>,
-    agent_session_id: Option<String>,
 }
 
 #[derive(Default)]
@@ -653,689 +649,7 @@ impl AcpxIdentifiers {
     }
 }
 
-fn resolve_profile(config: &LoongClawConfig) -> CliResult<ResolvedAcpxProfile> {
-    let profile = config.acp.acpx_profile().cloned().unwrap_or_default();
-    let command = profile
-        .command()
-        .unwrap_or_else(|| ACPX_DEFAULT_COMMAND.to_owned());
-    let cwd = profile.cwd();
-    let permission_mode = profile
-        .permission_mode()
-        .unwrap_or_else(|| ACPX_DEFAULT_PERMISSION_MODE.to_owned());
-    let non_interactive_permissions = profile
-        .non_interactive_permissions()
-        .unwrap_or_else(|| ACPX_DEFAULT_NON_INTERACTIVE_PERMISSIONS.to_owned());
-    let timeout_seconds = profile.timeout_seconds;
-    let queue_owner_ttl_seconds = profile
-        .queue_owner_ttl_seconds
-        .unwrap_or(ACPX_DEFAULT_QUEUE_OWNER_TTL_SECONDS);
-
-    if !matches!(
-        permission_mode.as_str(),
-        "approve-all" | "approve-reads" | "deny-all"
-    ) {
-        return Err(format!(
-            "ACPX permission_mode must be one of: approve-all, approve-reads, deny-all (got `{permission_mode}`)"
-        ));
-    }
-    if !matches!(non_interactive_permissions.as_str(), "deny" | "fail") {
-        return Err(format!(
-            "ACPX non_interactive_permissions must be one of: deny, fail (got `{non_interactive_permissions}`)"
-        ));
-    }
-    if timeout_seconds.is_some_and(|value| !value.is_finite() || value <= 0.0) {
-        return Err("ACPX timeout_seconds must be a positive finite number".to_owned());
-    }
-    if !queue_owner_ttl_seconds.is_finite() || queue_owner_ttl_seconds < 0.0 {
-        return Err("ACPX queue_owner_ttl_seconds must be a non-negative finite number".to_owned());
-    }
-
-    Ok(ResolvedAcpxProfile {
-        command,
-        cwd,
-        permission_mode,
-        non_interactive_permissions,
-        timeout_seconds,
-        queue_owner_ttl_seconds,
-        mcp_servers: profile.mcp_servers,
-    })
-}
-
-fn validate_requested_mcp_servers(
-    config: &LoongClawConfig,
-    profile: &ResolvedAcpxProfile,
-    request: &AcpSessionBootstrap,
-) -> CliResult<Vec<String>> {
-    if request.mcp_servers.is_empty() {
-        return Ok(Vec::new());
-    }
-    if !config.acp.allow_mcp_server_injection {
-        return Err(
-            "ACPX bootstrap requested MCP server injection but acp.allow_mcp_server_injection=false"
-                .to_owned(),
-        );
-    }
-
-    let mut selected = Vec::new();
-    let mut seen = BTreeSet::new();
-    let mut missing = Vec::new();
-    for raw_name in &request.mcp_servers {
-        let Some(name) = normalized_non_empty(raw_name.as_str()) else {
-            return Err("ACPX bootstrap mcp_servers entries must not be empty".to_owned());
-        };
-        if !profile.mcp_servers.contains_key(&name) {
-            missing.push(name);
-            continue;
-        }
-        if seen.insert(name.clone()) {
-            selected.push(name);
-        }
-    }
-
-    if missing.is_empty() {
-        Ok(selected)
-    } else {
-        Err(format!(
-            "ACPX requested mcp_servers are not configured under [acp.backends.acpx.mcp_servers]: {}",
-            missing.join(", ")
-        ))
-    }
-}
-
-async fn build_verb_args<I>(
-    profile: &ResolvedAcpxProfile,
-    timeout_ms: u64,
-    agent: &str,
-    cwd: &str,
-    selected_mcp_servers: &[String],
-    mut prefix: Vec<String>,
-    command: I,
-) -> CliResult<Vec<String>>
-where
-    I: IntoIterator<Item = String>,
-{
-    let raw_agent_command =
-        resolve_raw_agent_command(profile, timeout_ms, agent, cwd, selected_mcp_servers).await?;
-    if let Some(agent_command) = raw_agent_command {
-        prefix.extend(["--agent".to_owned(), agent_command]);
-    } else {
-        prefix.push(agent.to_owned());
-    }
-    prefix.extend(command);
-    Ok(prefix)
-}
-
-async fn build_prompt_args(
-    profile: &ResolvedAcpxProfile,
-    timeout_ms: u64,
-    agent: &str,
-    cwd: &str,
-    selected_mcp_servers: &[String],
-) -> CliResult<Vec<String>> {
-    let mut prompt_prefix = build_control_args(cwd);
-    prompt_prefix.extend(build_permission_args(profile.permission_mode.as_str()));
-    prompt_prefix.extend([
-        "--non-interactive-permissions".to_owned(),
-        profile.non_interactive_permissions.clone(),
-    ]);
-    if let Some(timeout_seconds) = profile.timeout_seconds {
-        prompt_prefix.extend(["--timeout".to_owned(), format_number(timeout_seconds)]);
-    }
-    prompt_prefix.extend([
-        "--ttl".to_owned(),
-        format_number(profile.queue_owner_ttl_seconds),
-    ]);
-
-    build_verb_args(
-        profile,
-        timeout_ms,
-        agent,
-        cwd,
-        selected_mcp_servers,
-        prompt_prefix,
-        Vec::<String>::new(),
-    )
-    .await
-}
-
-async fn resolve_raw_agent_command(
-    profile: &ResolvedAcpxProfile,
-    timeout_ms: u64,
-    agent: &str,
-    cwd: &str,
-    selected_mcp_servers: &[String],
-) -> CliResult<Option<String>> {
-    if selected_mcp_servers.is_empty() {
-        return Ok(None);
-    }
-
-    let target_command = resolve_acpx_agent_command(profile, timeout_ms, cwd, agent).await?;
-    let mcp_servers = resolve_selected_mcp_server_entries(profile, selected_mcp_servers)?;
-    let proxy_command = build_mcp_proxy_agent_command(target_command.as_str(), &mcp_servers)?;
-    Ok(Some(proxy_command))
-}
-
-async fn resolve_acpx_agent_command(
-    profile: &ResolvedAcpxProfile,
-    timeout_ms: u64,
-    cwd: &str,
-    agent: &str,
-) -> CliResult<String> {
-    let normalized_agent = agent.trim().to_ascii_lowercase();
-    let overrides = load_agent_overrides(profile, timeout_ms, cwd).await;
-    Ok(overrides
-        .get(&normalized_agent)
-        .cloned()
-        .or_else(|| builtin_agent_command(normalized_agent.as_str()))
-        .unwrap_or_else(|| agent.to_owned()))
-}
-
-async fn load_agent_overrides(
-    profile: &ResolvedAcpxProfile,
-    timeout_ms: u64,
-    cwd: &str,
-) -> BTreeMap<String, String> {
-    let args = vec![
-        "--cwd".to_owned(),
-        cwd.to_owned(),
-        "config".to_owned(),
-        "show".to_owned(),
-    ];
-    let Ok(output) = run_process(profile.command.as_str(), &args, cwd, timeout_ms, None).await
-    else {
-        return BTreeMap::new();
-    };
-    if output.exit_code.is_some_and(|code| code != 0) {
-        return BTreeMap::new();
-    }
-
-    let Ok(parsed) = serde_json::from_str::<Value>(output.stdout.as_str()) else {
-        return BTreeMap::new();
-    };
-    parsed
-        .get("agents")
-        .and_then(Value::as_object)
-        .map(|agents| {
-            agents
-                .iter()
-                .filter_map(|(name, entry)| {
-                    entry
-                        .get("command")
-                        .and_then(Value::as_str)
-                        .and_then(normalized_non_empty)
-                        .map(|command| (name.trim().to_ascii_lowercase(), command))
-                })
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn builtin_agent_command(agent: &str) -> Option<String> {
-    let command = match agent {
-        "codex" => "npx @zed-industries/codex-acp",
-        "claude" => "npx -y @zed-industries/claude-agent-acp",
-        "gemini" => "gemini",
-        "opencode" => "npx -y opencode-ai acp",
-        "pi" => "npx pi-acp",
-        _ => return None,
-    };
-    Some(command.to_owned())
-}
-
-fn resolve_selected_mcp_server_entries(
-    profile: &ResolvedAcpxProfile,
-    selected_mcp_servers: &[String],
-) -> CliResult<Vec<AcpxMcpServerEntry>> {
-    selected_mcp_servers
-        .iter()
-        .map(|name| {
-            let server = profile.mcp_servers.get(name).ok_or_else(|| {
-                format!(
-                    "ACPX requested mcp_servers are not configured under [acp.backends.acpx.mcp_servers]: {name}"
-                )
-            })?;
-            Ok(AcpxMcpServerEntry {
-                name: name.clone(),
-                command: server.command.clone(),
-                args: server.args.clone(),
-                cwd: None,
-                env: server
-                    .env
-                    .iter()
-                    .map(|(key, value)| AcpxMcpServerEnvEntry {
-                        name: key.clone(),
-                        value: value.clone(),
-                    })
-                    .collect(),
-            })
-        })
-        .collect()
-}
-
-fn resolve_effective_cwd(
-    request_cwd: Option<&PathBuf>,
-    profile_cwd: Option<&str>,
-) -> CliResult<String> {
-    if let Some(path) = request_cwd {
-        return Ok(path.display().to_string());
-    }
-    if let Some(cwd) = profile_cwd {
-        return Ok(cwd.to_owned());
-    }
-    std::env::current_dir()
-        .map(|path| path.display().to_string())
-        .map_err(|error| format!("resolve current working directory for ACPX failed: {error}"))
-}
-
-fn derive_agent_id(
-    config: &LoongClawConfig,
-    session_key: &str,
-    metadata: &BTreeMap<String, String>,
-) -> CliResult<String> {
-    let metadata_agent = metadata
-        .get("acp_agent")
-        .or_else(|| metadata.get("agent"))
-        .and_then(|value| normalized_non_empty(value));
-    let session_agent = parse_session_key_agent_id(session_key);
-
-    if let Some(session_agent) = session_agent {
-        let resolved = config.acp.resolve_allowed_agent(session_agent.as_str())?;
-        if let Some(metadata_agent) = metadata_agent {
-            let metadata_resolved = config.acp.resolve_allowed_agent(metadata_agent.as_str())?;
-            if metadata_resolved != resolved {
-                return Err(format!(
-                    "ACPX agent metadata `{metadata_resolved}` does not match session-key agent `{resolved}`"
-                ));
-            }
-        }
-        return Ok(resolved);
-    }
-
-    if let Some(metadata_agent) = metadata_agent {
-        return config.acp.resolve_allowed_agent(metadata_agent.as_str());
-    }
-
-    config.acp.resolved_default_agent()
-}
-
-fn parse_session_key_agent_id(session_key: &str) -> Option<String> {
-    session_key
-        .strip_prefix("agent:")
-        .and_then(|remainder| remainder.split_once(':').map(|(agent, _rest)| agent.trim()))
-        .filter(|agent| !agent.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn build_control_args(cwd: &str) -> Vec<String> {
-    vec![
-        "--format".to_owned(),
-        "json".to_owned(),
-        "--json-strict".to_owned(),
-        "--cwd".to_owned(),
-        cwd.to_owned(),
-    ]
-}
-
-fn build_permission_args(mode: &str) -> Vec<String> {
-    match mode {
-        "approve-all" => vec!["--approve-all".to_owned()],
-        "deny-all" => vec!["--deny-all".to_owned()],
-        _ => vec!["--approve-reads".to_owned()],
-    }
-}
-
-fn mode_label(mode: AcpSessionMode) -> &'static str {
-    match mode {
-        AcpSessionMode::Interactive => "interactive",
-        AcpSessionMode::Background => "background",
-        AcpSessionMode::Review => "review",
-    }
-}
-
-fn encode_runtime_handle_state(state: &AcpxRuntimeHandleState) -> CliResult<String> {
-    let payload = serde_json::to_vec(state)
-        .map_err(|error| format!("serialize ACPX runtime handle state failed: {error}"))?;
-    Ok(format!(
-        "{ACPX_HANDLE_PREFIX}{}",
-        base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(payload)
-    ))
-}
-
-fn decode_runtime_handle_state(
-    runtime_session_name: &str,
-) -> CliResult<Option<AcpxRuntimeHandleState>> {
-    let trimmed = runtime_session_name.trim();
-    let Some(encoded) = trimmed.strip_prefix(ACPX_HANDLE_PREFIX) else {
-        return Ok(None);
-    };
-    let decoded = base64::engine::general_purpose::URL_SAFE_NO_PAD
-        .decode(encoded)
-        .map_err(|error| format!("decode ACPX runtime handle state failed: {error}"))?;
-    serde_json::from_slice::<AcpxRuntimeHandleState>(&decoded)
-        .map(Some)
-        .map_err(|error| format!("parse ACPX runtime handle state failed: {error}"))
-}
-
-fn resolve_handle_state(
-    profile: &ResolvedAcpxProfile,
-    session: &AcpSessionHandle,
-) -> CliResult<AcpxRuntimeHandleState> {
-    if let Some(state) = decode_runtime_handle_state(session.runtime_session_name.as_str())? {
-        return Ok(state);
-    }
-
-    let cwd = session
-        .working_directory
-        .as_ref()
-        .map(|path| path.display().to_string())
-        .or_else(|| profile.cwd.clone())
-        .map(Ok)
-        .unwrap_or_else(|| {
-            std::env::current_dir()
-                .map(|path| path.display().to_string())
-                .map_err(|error| {
-                    format!("resolve current working directory for ACPX handle failed: {error}")
-                })
-        })?;
-    let name = normalized_non_empty(session.runtime_session_name.as_str())
-        .unwrap_or_else(|| session.session_key.clone());
-
-    Ok(AcpxRuntimeHandleState {
-        name,
-        agent: parse_session_key_agent_id(session.session_key.as_str())
-            .unwrap_or_else(|| ACPX_DEFAULT_AGENT.to_owned()),
-        cwd,
-        mode: "persistent".to_owned(),
-        mcp_servers: Vec::new(),
-        acpx_record_id: None,
-        backend_session_id: session.backend_session_id.clone(),
-        agent_session_id: session.agent_session_id.clone(),
-    })
-}
-
-async fn run_json_command(
-    profile: &ResolvedAcpxProfile,
-    args: Vec<String>,
-    cwd: &str,
-    timeout_ms: u64,
-    stdin_payload: Option<&str>,
-    ignore_no_session: bool,
-) -> CliResult<Vec<Value>> {
-    let output = run_process(
-        profile.command.as_str(),
-        &args,
-        cwd,
-        timeout_ms,
-        stdin_payload,
-    )
-    .await?;
-    let events = parse_json_lines(output.stdout.as_str());
-    if let Some(error) = event_error_message(&events, ignore_no_session) {
-        return Err(error);
-    }
-    if output.exit_code.is_some_and(|code| code != 0) {
-        return Err(format_exit_message(
-            output.stderr.as_str(),
-            output.exit_code,
-        ));
-    }
-    Ok(events)
-}
-
-async fn run_prompt_process(
-    command: &str,
-    args: &[String],
-    cwd: &str,
-    timeout_ms: u64,
-    stdin_payload: &str,
-    mut abort: Option<AcpAbortSignal>,
-    sink: Option<&dyn AcpTurnEventSink>,
-) -> CliResult<AcpTurnResult> {
-    if !Path::new(cwd).exists() {
-        return Err(format!(
-            "ACP runtime working directory does not exist: {cwd}"
-        ));
-    }
-
-    if abort.as_ref().is_some_and(AcpAbortSignal::is_aborted) {
-        let done = synthetic_done_event(Some(AcpTurnStopReason::Cancelled));
-        emit_turn_event(sink, &done)?;
-        return Ok(AcpTurnResult {
-            output_text: String::new(),
-            state: AcpSessionState::Ready,
-            usage: None,
-            events: vec![done],
-            stop_reason: Some(AcpTurnStopReason::Cancelled),
-        });
-    }
-
-    let mut child = spawn_acpx_child(command, args, cwd, true).await?;
-
-    let Some(mut stdin) = child.stdin.take() else {
-        return Err("ACPX command spawned without stdin pipe".to_owned());
-    };
-    stdin
-        .write_all(stdin_payload.as_bytes())
-        .await
-        .map_err(|error| format!("write ACPX stdin failed: {error}"))?;
-    drop(stdin);
-
-    let Some(stdout) = child.stdout.take() else {
-        return Err("ACPX command spawned without stdout pipe".to_owned());
-    };
-    let Some(child_stderr) = child.stderr.take() else {
-        return Err("ACPX command spawned without stderr pipe".to_owned());
-    };
-
-    let mut stderr_task = Some(tokio::spawn(async move {
-        let mut stderr = String::new();
-        let mut reader = BufReader::new(child_stderr);
-        reader
-            .read_to_string(&mut stderr)
-            .await
-            .map_err(|error| format!("read ACPX stderr failed: {error}"))?;
-        Ok::<String, String>(stderr)
-    }));
-
-    let deadline = Instant::now() + Duration::from_millis(timeout_ms);
-    let mut lines = BufReader::new(stdout).lines();
-    let mut events = Vec::new();
-    let mut saw_done = false;
-    let mut saw_error = false;
-
-    loop {
-        let abort_enabled = abort.is_some();
-        tokio::select! {
-            _ = sleep_until(deadline) => {
-                terminate_child_process(&mut child).await;
-                abort_stderr_task(&mut stderr_task);
-                return Err(format!(
-                    "ACPX command timed out after {timeout_ms}ms: {} {}",
-                    command,
-                    args.join(" ")
-                ));
-            }
-            _ = wait_for_abort(&mut abort), if abort_enabled => {
-                terminate_child_process(&mut child).await;
-                abort_stderr_task(&mut stderr_task);
-                let done = synthetic_done_event(Some(AcpTurnStopReason::Cancelled));
-                emit_turn_event(sink, &done)?;
-                events.push(done);
-                return Ok(AcpTurnResult {
-                    output_text: collect_output_text(&events),
-                    state: AcpSessionState::Ready,
-                    usage: collect_usage_update(&events),
-                    stop_reason: Some(AcpTurnStopReason::Cancelled),
-                    events,
-                });
-            }
-            line = lines.next_line() => {
-                let line = line.map_err(|error| format!("read ACPX stdout failed: {error}"))?;
-                let Some(line) = line else {
-                    break;
-                };
-                let Some(event) = parse_json_line(line.as_str()) else {
-                    continue;
-                };
-                saw_done |= is_done_event(&event);
-                saw_error |= value_string(&event, "type").as_deref() == Some("error");
-                emit_turn_event(sink, &event)?;
-                events.push(event);
-            }
-        }
-    }
-
-    let abort_enabled = abort.is_some();
-    let exit_status = tokio::select! {
-        _ = sleep_until(deadline) => {
-            terminate_child_process(&mut child).await;
-            abort_stderr_task(&mut stderr_task);
-            return Err(format!(
-                "ACPX command timed out after {timeout_ms}ms: {} {}",
-                command,
-                args.join(" ")
-            ));
-        }
-        _ = wait_for_abort(&mut abort), if abort_enabled => {
-            terminate_child_process(&mut child).await;
-            abort_stderr_task(&mut stderr_task);
-            let done = synthetic_done_event(Some(AcpTurnStopReason::Cancelled));
-            emit_turn_event(sink, &done)?;
-            events.push(done);
-            return Ok(AcpTurnResult {
-                output_text: collect_output_text(&events),
-                state: AcpSessionState::Ready,
-                usage: collect_usage_update(&events),
-                stop_reason: Some(AcpTurnStopReason::Cancelled),
-                events,
-            });
-        }
-        status = child.wait() => {
-            status.map_err(|error| format!("wait for ACPX command failed: {error}"))?
-        }
-    };
-    let stderr = collect_stderr_task(&mut stderr_task).await?;
-
-    if let Some(error) = event_error_message(&events, false) {
-        return Err(error);
-    }
-    if exit_status.code().is_some_and(|code| code != 0) {
-        return Err(format_exit_message(stderr.as_str(), exit_status.code()));
-    }
-    if !saw_done && !saw_error {
-        let done = synthetic_done_event(Some(AcpTurnStopReason::Completed));
-        emit_turn_event(sink, &done)?;
-        events.push(done);
-    }
-
-    Ok(AcpTurnResult {
-        output_text: collect_output_text(&events),
-        state: AcpSessionState::Ready,
-        usage: collect_usage_update(&events),
-        stop_reason: collect_stop_reason(&events).or(Some(AcpTurnStopReason::Completed)),
-        events,
-    })
-}
-
-async fn run_process(
-    command: &str,
-    args: &[String],
-    cwd: &str,
-    timeout_ms: u64,
-    stdin_payload: Option<&str>,
-) -> CliResult<AcpxCommandOutput> {
-    if !Path::new(cwd).exists() {
-        return Err(format!(
-            "ACP runtime working directory does not exist: {cwd}"
-        ));
-    }
-
-    let mut child = spawn_acpx_child(command, args, cwd, stdin_payload.is_some()).await?;
-
-    if let Some(payload) = stdin_payload {
-        let Some(mut stdin) = child.stdin.take() else {
-            return Err("ACPX command spawned without stdin pipe".to_owned());
-        };
-        stdin
-            .write_all(payload.as_bytes())
-            .await
-            .map_err(|error| format!("write ACPX stdin failed: {error}"))?;
-        drop(stdin);
-    }
-
-    let output = timeout(Duration::from_millis(timeout_ms), child.wait_with_output())
-        .await
-        .map_err(|error| {
-            format!(
-                "ACPX command timed out after {timeout_ms}ms: {} {} ({error})",
-                command,
-                args.join(" ")
-            )
-        })
-        .and_then(|result| {
-            result.map_err(|error| format!("wait for ACPX command failed: {error}"))
-        })?;
-
-    Ok(AcpxCommandOutput {
-        stdout: String::from_utf8_lossy(&output.stdout).to_string(),
-        stderr: String::from_utf8_lossy(&output.stderr).to_string(),
-        exit_code: output.status.code(),
-    })
-}
-
-async fn collect_stderr_task(
-    task: &mut Option<tokio::task::JoinHandle<Result<String, String>>>,
-) -> CliResult<String> {
-    let Some(task) = task.take() else {
-        return Ok(String::new());
-    };
-    match task.await {
-        Ok(Ok(stderr)) => Ok(stderr),
-        Ok(Err(error)) => Err(error),
-        Err(error) => Err(format!("join ACPX stderr reader failed: {error}")),
-    }
-}
-
-fn abort_stderr_task(task: &mut Option<tokio::task::JoinHandle<Result<String, String>>>) {
-    if let Some(task) = task.take() {
-        task.abort();
-    }
-}
-
-async fn terminate_child_process(child: &mut tokio::process::Child) {
-    let _ = child.start_kill();
-    let _ = child.wait().await;
-}
-
-async fn wait_for_abort(abort: &mut Option<AcpAbortSignal>) {
-    if let Some(abort) = abort.as_mut() {
-        abort.cancelled().await;
-    } else {
-        std::future::pending::<()>().await;
-    }
-}
-
-fn synthetic_done_event(stop_reason: Option<AcpTurnStopReason>) -> Value {
-    let mut event = serde_json::Map::from_iter([("type".to_owned(), json!("done"))]);
-    if let Some(stop_reason) = stop_reason {
-        event.insert(
-            "stopReason".to_owned(),
-            json!(match stop_reason {
-                AcpTurnStopReason::Completed => "completed",
-                AcpTurnStopReason::Cancelled => "cancelled",
-            }),
-        );
-    }
-    Value::Object(event)
-}
-
-fn emit_turn_event(sink: Option<&dyn AcpTurnEventSink>, event: &Value) -> CliResult<()> {
-    if let Some(sink) = sink {
-        sink.on_event(event)?;
-    }
-    Ok(())
-}
-
+#[allow(dead_code)]
 fn map_spawn_error(command: &str, cwd: &str, error: std::io::Error) -> String {
     if error.kind() == ErrorKind::NotFound {
         if !Path::new(cwd).exists() {
@@ -1346,6 +660,7 @@ fn map_spawn_error(command: &str, cwd: &str, error: std::io::Error) -> String {
     format!("spawn ACPX command failed: {error}")
 }
 
+#[allow(dead_code)]
 async fn spawn_acpx_child(
     command: &str,
     args: &[String],
@@ -1370,6 +685,7 @@ async fn spawn_acpx_child(
     .map_err(|error| map_spawn_error(command, cwd, error))
 }
 
+#[cfg(test)]
 async fn retry_executable_file_busy<T, F>(operation: F) -> std::io::Result<T>
 where
     F: FnMut() -> std::io::Result<T>,
@@ -1385,188 +701,6 @@ where
     F: FnMut() -> std::io::Result<T>,
 {
     retry_spawn_blocking(operation, ACPX_SPAWN_RETRY_ATTEMPTS, ACPX_SPAWN_RETRY_DELAY)
-}
-
-fn parse_json_lines(stdout: &str) -> Vec<Value> {
-    stdout.lines().filter_map(parse_json_line).collect()
-}
-
-fn parse_json_line(line: &str) -> Option<Value> {
-    let trimmed = line.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-    serde_json::from_str::<Value>(trimmed)
-        .ok()
-        .filter(|value| value.is_object())
-}
-
-fn is_done_event(event: &Value) -> bool {
-    value_string(event, "type").as_deref() == Some("done")
-}
-
-fn extract_identifiers(events: &[Value]) -> AcpxIdentifiers {
-    let mut identifiers = AcpxIdentifiers::default();
-    for event in events {
-        if identifiers.acpx_record_id.is_none() {
-            identifiers.acpx_record_id = value_string(event, "acpxRecordId");
-        }
-        if identifiers.backend_session_id.is_none() {
-            identifiers.backend_session_id = value_string(event, "acpxSessionId")
-                .or_else(|| value_string(event, "backendSessionId"));
-        }
-        if identifiers.agent_session_id.is_none() {
-            identifiers.agent_session_id = value_string(event, "agentSessionId");
-        }
-    }
-    identifiers
-}
-
-fn collect_output_text(events: &[Value]) -> String {
-    let mut output = String::new();
-    for event in events {
-        if let Some(chunk) = extract_output_chunk(event) {
-            output.push_str(chunk.as_str());
-        }
-    }
-    output
-}
-
-fn extract_output_chunk(event: &Value) -> Option<String> {
-    if let Some(kind) = value_string(event, "type") {
-        if kind == "text" {
-            return raw_string(event, "content");
-        }
-        if kind == "agent_message_chunk" {
-            return nested_text(event);
-        }
-    }
-
-    if value_string(event, "sessionUpdate").as_deref() == Some("agent_message_chunk") {
-        return nested_text(event);
-    }
-
-    let payload = event
-        .get("params")
-        .and_then(|params| params.get("update"))?;
-    if payload.get("sessionUpdate").and_then(Value::as_str) == Some("agent_message_chunk") {
-        return nested_text(payload);
-    }
-    None
-}
-
-fn collect_usage_update(events: &[Value]) -> Option<Value> {
-    events.iter().rev().find_map(|event| {
-        let direct_usage = value_string(event, "type").as_deref() == Some("usage_update");
-        let tagged_usage = value_string(event, "sessionUpdate").as_deref() == Some("usage_update");
-        let nested_usage = event
-            .get("params")
-            .and_then(|params| params.get("update"))
-            .and_then(|payload| payload.get("sessionUpdate"))
-            .and_then(Value::as_str)
-            == Some("usage_update");
-        if !(direct_usage || tagged_usage || nested_usage) {
-            return None;
-        }
-
-        let payload = event
-            .get("params")
-            .and_then(|params| params.get("update"))
-            .unwrap_or(event);
-        let used = payload.get("used").and_then(Value::as_u64);
-        let size = payload.get("size").and_then(Value::as_u64);
-        if used.is_none() && size.is_none() {
-            return None;
-        }
-
-        let mut usage = serde_json::Map::new();
-        if let Some(used) = used {
-            usage.insert("used".to_owned(), json!(used));
-        }
-        if let Some(size) = size {
-            usage.insert("size".to_owned(), json!(size));
-        }
-        Some(Value::Object(usage))
-    })
-}
-
-fn collect_stop_reason(events: &[Value]) -> Option<AcpTurnStopReason> {
-    events.iter().rev().find_map(|event| {
-        let reason =
-            value_string(event, "stopReason").or_else(|| value_string(event, "stop_reason"))?;
-        match reason.to_ascii_lowercase().as_str() {
-            "cancel" | "cancelled" => Some(AcpTurnStopReason::Cancelled),
-            "complete" | "completed" | "done" => Some(AcpTurnStopReason::Completed),
-            _ => None,
-        }
-    })
-}
-
-fn nested_text(value: &Value) -> Option<String> {
-    if let Some(text) = raw_string(value, "text") {
-        return Some(text);
-    }
-    let content = value.get("content")?;
-    match content {
-        Value::String(text) => (!text.is_empty()).then(|| text.clone()),
-        Value::Object(_) => raw_string(content, "text"),
-        Value::Null | Value::Bool(_) | Value::Number(_) | Value::Array(_) => None,
-    }
-}
-
-fn raw_string(value: &Value, key: &str) -> Option<String> {
-    value
-        .get(key)
-        .and_then(Value::as_str)
-        .map(ToOwned::to_owned)
-        .filter(|text| !text.is_empty())
-}
-
-fn value_string(value: &Value, key: &str) -> Option<String> {
-    value
-        .get(key)
-        .and_then(Value::as_str)
-        .and_then(normalized_non_empty)
-}
-
-fn normalized_non_empty(value: &str) -> Option<String> {
-    let trimmed = value.trim();
-    (!trimmed.is_empty()).then(|| trimmed.to_owned())
-}
-
-fn event_error_message(events: &[Value], ignore_no_session: bool) -> Option<String> {
-    let code = event_code(events);
-    let message = events.iter().find_map(|event| {
-        (value_string(event, "type").as_deref() == Some("error"))
-            .then(|| value_string(event, "message"))
-            .flatten()
-    })?;
-    if ignore_no_session && code.as_deref() == Some("NO_SESSION") {
-        return None;
-    }
-    Some(match code {
-        Some(code) => format!("{code}: {message}"),
-        None => message,
-    })
-}
-
-fn event_code(events: &[Value]) -> Option<String> {
-    events.iter().find_map(|event| {
-        (value_string(event, "type").as_deref() == Some("error"))
-            .then(|| value_string(event, "code"))
-            .flatten()
-    })
-}
-
-fn map_status_state(raw: Option<&str>) -> AcpSessionState {
-    match raw.map(str::trim).map(str::to_ascii_lowercase).as_deref() {
-        Some("busy") | Some("running") | Some("active") => AcpSessionState::Busy,
-        Some("cancelling") | Some("cancelled") => AcpSessionState::Cancelling,
-        Some("error") | Some("failed") => AcpSessionState::Error,
-        Some("closed") | Some("terminated") => AcpSessionState::Closed,
-        Some("initializing") | Some("starting") => AcpSessionState::Initializing,
-        _ => AcpSessionState::Ready,
-    }
 }
 
 fn format_exit_message(stderr: &str, exit_code: Option<i32>) -> String {
