+++ crates/app/src/acp/manager.rs
@@ -1,12 +1,23 @@
 use std::collections::BTreeMap;
 use std::sync::{Arc, RwLock};
 
-use sha2::{Digest, Sha256};
 use tokio::sync::{Mutex as AsyncMutex, OwnedMutexGuard};
 
 use crate::CliResult;
 use crate::config::LoongClawConfig;
 
+#[path = "manager_actor.rs"]
+mod actor;
+#[path = "manager_sessions.rs"]
+mod sessions;
+#[path = "manager_support.rs"]
+mod support;
+
+use self::support::*;
+pub use self::support::{
+    AcpManagerActorSnapshot, AcpManagerObservabilitySnapshot, AcpManagerRuntimeCacheSnapshot,
+    AcpManagerSessionSnapshot, AcpManagerTurnSnapshot,
+};
 use super::backend::{
     ACP_SESSION_METADATA_ACTIVATION_ORIGIN, ACP_TURN_METADATA_TRACE_ID, AcpAbortController,
     AcpConfigPatch, AcpDoctorReport, AcpRoutingOrigin, AcpSessionBootstrap, AcpSessionHandle,
@@ -18,56 +29,6 @@ use super::merge_turn_events;
 use super::registry::resolve_acp_backend;
 use super::runtime::resolve_acp_backend_selection;
 use super::store::{AcpSessionStore, InMemoryAcpSessionStore};
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AcpManagerRuntimeCacheSnapshot {
-    pub active_sessions: usize,
-    pub idle_ttl_ms: u64,
-    pub evicted_total: u64,
-    pub last_evicted_at_ms: Option<u64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AcpManagerTurnSnapshot {
-    pub active: usize,
-    pub queue_depth: usize,
-    pub completed: u64,
-    pub failed: u64,
-    pub average_latency_ms: u64,
-    pub max_latency_ms: u64,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AcpManagerActorSnapshot {
-    pub active: usize,
-    pub queue_depth: usize,
-    pub waiting: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AcpManagerSessionSnapshot {
-    pub bound: usize,
-    pub unbound: usize,
-    pub activation_origin_counts: BTreeMap<String, usize>,
-    pub backend_counts: BTreeMap<String, usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AcpManagerObservabilitySnapshot {
-    pub runtime_cache: AcpManagerRuntimeCacheSnapshot,
-    pub sessions: AcpManagerSessionSnapshot,
-    pub actors: AcpManagerActorSnapshot,
-    pub turns: AcpManagerTurnSnapshot,
-    pub errors_by_code: BTreeMap<String, usize>,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-struct TurnLatencyStats {
-    completed: u64,
-    failed: u64,
-    total_ms: u64,
-    max_ms: u64,
-}
 
 #[derive(Debug)]
 struct ActiveTurnState {
@@ -649,572 +610,6 @@ impl AcpSessionManager {
     pub fn list_sessions(&self) -> CliResult<Vec<AcpSessionMetadata>> {
         self.store.list()
     }
-
-    fn resolve_existing_session(
-        &self,
-        config: &LoongClawConfig,
-        selected_backend_id: &str,
-        bootstrap: &AcpSessionBootstrap,
-    ) -> CliResult<Option<AcpSessionMetadata>> {
-        if let Some(existing) = self.store.get(bootstrap.session_key.as_str())? {
-            return self
-                .validate_and_touch_existing_session(selected_backend_id, bootstrap, existing)
-                .map(Some);
-        }
-
-        if config.acp.bindings_enabled {
-            if let Some(binding) = AcpSessionBindingScope::from_bootstrap(bootstrap)
-                && let Some(existing) = self
-                    .store
-                    .get_by_binding_route_session_id(binding.route_session_id.as_str())?
-            {
-                return self
-                    .validate_and_touch_existing_session(selected_backend_id, bootstrap, existing)
-                    .map(Some);
-            }
-            if let Some(conversation_id) =
-                normalized_conversation_id(bootstrap.conversation_id.as_deref())
-                && let Some(existing) = self
-                    .store
-                    .get_by_conversation_id(conversation_id.as_str())?
-            {
-                return self
-                    .validate_and_touch_existing_session(selected_backend_id, bootstrap, existing)
-                    .map(Some);
-            }
-        }
-
-        Ok(None)
-    }
-
-    fn validate_and_touch_existing_session(
-        &self,
-        selected_backend_id: &str,
-        bootstrap: &AcpSessionBootstrap,
-        mut existing: AcpSessionMetadata,
-    ) -> CliResult<AcpSessionMetadata> {
-        if existing.backend_id != selected_backend_id {
-            return Err(format!(
-                "session `{}` is already bound to ACP backend `{}` (requested `{}`); use a new session key or close the existing session first",
-                existing.session_key, existing.backend_id, selected_backend_id
-            ));
-        }
-
-        if let Some(binding) = AcpSessionBindingScope::from_bootstrap(bootstrap) {
-            if let Some(existing_binding) = existing.binding.as_ref() {
-                if existing_binding != &binding {
-                    return Err(format!(
-                        "session `{}` is already bound to ACP route `{}` (requested `{}`); use a new session key or close the existing session first",
-                        existing.session_key,
-                        existing_binding.route_session_id,
-                        binding.route_session_id
-                    ));
-                }
-            } else {
-                existing.binding = Some(binding);
-            }
-        }
-
-        if let Some(conversation_id) =
-            normalized_conversation_id(bootstrap.conversation_id.as_deref())
-        {
-            if existing.binding.is_some() {
-                if existing.conversation_id.is_none() {
-                    existing.conversation_id = Some(conversation_id);
-                }
-            } else {
-                existing.conversation_id = Some(conversation_id);
-            }
-        }
-        if existing.mode.is_none() {
-            existing.mode = bootstrap.mode;
-        }
-        existing.touch();
-        self.store.upsert(existing.clone())?;
-        Ok(existing)
-    }
-
-    fn enforce_max_concurrent_sessions(
-        &self,
-        config: &LoongClawConfig,
-        requested_session_key: &str,
-    ) -> CliResult<()> {
-        if self.store.get(requested_session_key)?.is_some() {
-            return Ok(());
-        }
-        let current = self.store.list()?.len();
-        let limit = config.acp.max_concurrent_sessions();
-        if current >= limit {
-            return Err(format!(
-                "ACP control plane already tracks {current} sessions, which reaches max_concurrent_sessions={limit}"
-            ));
-        }
-        Ok(())
-    }
-
-    async fn cleanup_idle_sessions(&self, config: &LoongClawConfig) -> CliResult<()> {
-        let ttl_ms = config.acp.session_idle_ttl_ms();
-        if ttl_ms == 0 {
-            return Ok(());
-        }
-
-        let now = now_ms();
-        for metadata in self.store.list()? {
-            if matches!(
-                metadata.state,
-                AcpSessionState::Busy | AcpSessionState::Cancelling | AcpSessionState::Initializing
-            ) {
-                continue;
-            }
-            if self.actor_ref_count_for_metadata(&metadata)? > 0 {
-                continue;
-            }
-            if self.pending_turn_count_for_metadata(&metadata)? > 0 {
-                continue;
-            }
-            if self.is_active_turn_for_metadata(&metadata)? {
-                continue;
-            }
-            if now.saturating_sub(metadata.last_activity_ms) < ttl_ms {
-                continue;
-            }
-
-            if let Ok(backend) = resolve_acp_backend(Some(metadata.backend_id.as_str())) {
-                let _ = backend.close(config, &metadata.to_handle()).await;
-            }
-            let _ = self.store.remove(metadata.session_key.as_str());
-            self.record_eviction(now)?;
-        }
-
-        Ok(())
-    }
-
-    fn fallback_status(
-        &self,
-        metadata: &AcpSessionMetadata,
-        active_turn: bool,
-        pending_turns: usize,
-    ) -> AcpSessionStatus {
-        AcpSessionStatus {
-            session_key: metadata.session_key.clone(),
-            backend_id: metadata.backend_id.clone(),
-            conversation_id: metadata.conversation_id.clone(),
-            binding: metadata.binding.clone(),
-            activation_origin: metadata.activation_origin,
-            state: projected_status_state(metadata.state, active_turn, pending_turns),
-            mode: metadata.mode,
-            pending_turns: pending_turns.max(usize::from(active_turn)),
-            active_turn_id: active_turn.then(|| metadata.runtime_session_name.clone()),
-            last_activity_ms: metadata.last_activity_ms,
-            last_error: metadata.last_error.clone(),
-        }
-    }
-
-    async fn request_active_turn_cancellation(
-        &self,
-        config: &LoongClawConfig,
-        mut metadata: AcpSessionMetadata,
-        active_turn: Arc<ActiveTurnState>,
-    ) -> CliResult<()> {
-        active_turn.abort_controller.abort();
-
-        metadata.state = AcpSessionState::Cancelling;
-        metadata.clear_error();
-        metadata.touch();
-        self.store.upsert(metadata.clone())?;
-
-        let backend = resolve_acp_backend(Some(active_turn.handle.backend_id.as_str()))?;
-        match backend.cancel(config, &active_turn.handle).await {
-            Ok(()) => Ok(()),
-            Err(error) => {
-                self.record_error(error.as_str())?;
-                metadata.state = AcpSessionState::Error;
-                metadata.set_error(error.clone());
-                self.store.upsert(metadata)?;
-                Err(error)
-            }
-        }
-    }
-
-    async fn acquire_session_actor_guard(&self, actor_key: String) -> CliResult<SessionActorGuard> {
-        self.increment_actor_ref_count(actor_key.as_str())?;
-        let queue_lock = match self.get_or_insert_session_actor(actor_key.as_str()) {
-            Ok(lock) => lock,
-            Err(error) => {
-                let _ = self.decrement_actor_ref_count(actor_key.as_str());
-                return Err(error);
-            }
-        };
-
-        Ok(SessionActorGuard {
-            actor_key,
-            actor_guard: Some(queue_lock.lock_owned().await),
-            session_actor_locks: self.session_actor_locks.clone(),
-            actor_ref_counts: self.actor_ref_counts.clone(),
-            pending_turns: self.pending_turns.clone(),
-            count_pending_turn: false,
-        })
-    }
-
-    async fn acquire_turn_queue_guard(&self, actor_key: String) -> CliResult<SessionActorGuard> {
-        self.increment_actor_ref_count(actor_key.as_str())?;
-        if let Err(error) = self.increment_pending_turn(actor_key.as_str()) {
-            let _ = self.decrement_actor_ref_count(actor_key.as_str());
-            return Err(error);
-        }
-        let queue_lock = match self.get_or_insert_session_actor(actor_key.as_str()) {
-            Ok(lock) => lock,
-            Err(error) => {
-                let _ = self.decrement_pending_turn(actor_key.as_str());
-                let _ = self.decrement_actor_ref_count(actor_key.as_str());
-                return Err(error);
-            }
-        };
-
-        Ok(SessionActorGuard {
-            actor_key,
-            actor_guard: Some(queue_lock.lock_owned().await),
-            session_actor_locks: self.session_actor_locks.clone(),
-            actor_ref_counts: self.actor_ref_counts.clone(),
-            pending_turns: self.pending_turns.clone(),
-            count_pending_turn: true,
-        })
-    }
-
-    fn get_or_insert_session_actor(&self, actor_key: &str) -> CliResult<Arc<AsyncMutex<()>>> {
-        let mut guard = self
-            .session_actor_locks
-            .write()
-            .map_err(|_error| "ACP session actor registry lock poisoned".to_owned())?;
-        Ok(guard
-            .entry(actor_key.to_owned())
-            .or_insert_with(|| Arc::new(AsyncMutex::new(())))
-            .clone())
-    }
-
-    fn increment_actor_ref_count(&self, actor_key: &str) -> CliResult<()> {
-        let mut guard = self
-            .actor_ref_counts
-            .write()
-            .map_err(|_error| "ACP actor reference registry lock poisoned".to_owned())?;
-        *guard.entry(actor_key.to_owned()).or_insert(0) += 1;
-        Ok(())
-    }
-
-    fn decrement_actor_ref_count(&self, actor_key: &str) -> CliResult<()> {
-        let mut guard = self
-            .actor_ref_counts
-            .write()
-            .map_err(|_error| "ACP actor reference registry lock poisoned".to_owned())?;
-        if let Some(count) = guard.get_mut(actor_key) {
-            if *count <= 1 {
-                guard.remove(actor_key);
-            } else {
-                *count -= 1;
-            }
-        }
-        Ok(())
-    }
-
-    fn increment_pending_turn(&self, actor_key: &str) -> CliResult<()> {
-        let mut guard = self
-            .pending_turns
-            .write()
-            .map_err(|_error| "ACP pending turn registry lock poisoned".to_owned())?;
-        *guard.entry(actor_key.to_owned()).or_insert(0) += 1;
-        Ok(())
-    }
-
-    fn decrement_pending_turn(&self, actor_key: &str) -> CliResult<()> {
-        let mut guard = self
-            .pending_turns
-            .write()
-            .map_err(|_error| "ACP pending turn registry lock poisoned".to_owned())?;
-        if let Some(count) = guard.get_mut(actor_key) {
-            if *count <= 1 {
-                guard.remove(actor_key);
-            } else {
-                *count -= 1;
-            }
-        }
-        Ok(())
-    }
-
-    fn pending_turn_count_for_metadata(&self, metadata: &AcpSessionMetadata) -> CliResult<usize> {
-        self.pending_turn_count(actor_key_for_metadata(metadata).as_str())
-    }
-
-    fn actor_ref_count_for_metadata(&self, metadata: &AcpSessionMetadata) -> CliResult<usize> {
-        self.actor_ref_count(actor_key_for_metadata(metadata).as_str())
-    }
-
-    fn pending_turn_count(&self, actor_key: &str) -> CliResult<usize> {
-        let guard = self
-            .pending_turns
-            .read()
-            .map_err(|_error| "ACP pending turn registry lock poisoned".to_owned())?;
-        Ok(guard.get(actor_key).copied().unwrap_or(0))
-    }
-
-    fn actor_ref_count(&self, actor_key: &str) -> CliResult<usize> {
-        let guard = self
-            .actor_ref_counts
-            .read()
-            .map_err(|_error| "ACP actor reference registry lock poisoned".to_owned())?;
-        Ok(guard.get(actor_key).copied().unwrap_or(0))
-    }
-
-    fn active_turn(&self, actor_key: &str) -> CliResult<Option<Arc<ActiveTurnState>>> {
-        let guard = self
-            .active_turns
-            .read()
-            .map_err(|_error| "ACP active turn registry lock poisoned".to_owned())?;
-        Ok(guard.get(actor_key).cloned())
-    }
-
-    fn is_active_turn_for_metadata(&self, metadata: &AcpSessionMetadata) -> CliResult<bool> {
-        Ok(self
-            .active_turn(actor_key_for_metadata(metadata).as_str())?
-            .is_some())
-    }
-
-    fn record_turn_completion(&self, started_ms: u64, succeeded: bool) -> CliResult<()> {
-        let duration_ms = now_ms().saturating_sub(started_ms);
-        let mut guard = self
-            .turn_latency_stats
-            .write()
-            .map_err(|_error| "ACP turn latency registry lock poisoned".to_owned())?;
-        if succeeded {
-            guard.completed = guard.completed.saturating_add(1);
-        } else {
-            guard.failed = guard.failed.saturating_add(1);
-        }
-        guard.total_ms = guard.total_ms.saturating_add(duration_ms);
-        guard.max_ms = guard.max_ms.max(duration_ms);
-        Ok(())
-    }
-
-    fn record_error(&self, error: &str) -> CliResult<()> {
-        let key = normalize_error_key(error);
-        let mut guard = self
-            .error_counts_by_code
-            .write()
-            .map_err(|_error| "ACP error registry lock poisoned".to_owned())?;
-        *guard.entry(key).or_insert(0) += 1;
-        Ok(())
-    }
-
-    fn record_eviction(&self, at_ms: u64) -> CliResult<()> {
-        let mut count_guard = self
-            .evicted_runtime_count
-            .write()
-            .map_err(|_error| "ACP eviction counter lock poisoned".to_owned())?;
-        *count_guard = count_guard.saturating_add(1);
-        let mut ts_guard = self
-            .last_evicted_at_ms
-            .write()
-            .map_err(|_error| "ACP last eviction lock poisoned".to_owned())?;
-        *ts_guard = Some(at_ms);
-        Ok(())
-    }
-
-    fn register_active_turn(
-        &self,
-        actor_key: &str,
-        active_turn: Arc<ActiveTurnState>,
-    ) -> CliResult<()> {
-        let mut guard = self
-            .active_turns
-            .write()
-            .map_err(|_error| "ACP active turn registry lock poisoned".to_owned())?;
-        guard.insert(actor_key.to_owned(), active_turn);
-        Ok(())
-    }
-
-    fn clear_active_turn(&self, actor_key: &str) -> CliResult<()> {
-        let mut guard = self
-            .active_turns
-            .write()
-            .map_err(|_error| "ACP active turn registry lock poisoned".to_owned())?;
-        guard.remove(actor_key);
-        Ok(())
-    }
-}
-
-struct SessionActorGuard {
-    actor_key: String,
-    actor_guard: Option<OwnedMutexGuard<()>>,
-    session_actor_locks: Arc<RwLock<BTreeMap<String, Arc<AsyncMutex<()>>>>>,
-    actor_ref_counts: Arc<RwLock<BTreeMap<String, usize>>>,
-    pending_turns: Arc<RwLock<BTreeMap<String, usize>>>,
-    count_pending_turn: bool,
-}
-
-impl Drop for SessionActorGuard {
-    fn drop(&mut self) {
-        self.actor_guard.take();
-
-        if self.count_pending_turn
-            && let Ok(mut guard) = self.pending_turns.write()
-        {
-            match guard.get_mut(self.actor_key.as_str()) {
-                Some(count) if *count <= 1 => {
-                    guard.remove(self.actor_key.as_str());
-                }
-                Some(count) => {
-                    *count -= 1;
-                }
-                None => {}
-            }
-        }
-
-        if let Ok(mut guard) = self.actor_ref_counts.write() {
-            match guard.get_mut(self.actor_key.as_str()) {
-                Some(count) if *count <= 1 => {
-                    guard.remove(self.actor_key.as_str());
-                }
-                Some(count) => {
-                    *count -= 1;
-                }
-                None => {}
-            }
-        }
-
-        let should_remove_actor = self
-            .actor_ref_counts
-            .read()
-            .map(|guard| !guard.contains_key(self.actor_key.as_str()))
-            .unwrap_or(false);
-        if should_remove_actor && let Ok(mut guard) = self.session_actor_locks.write() {
-            guard.remove(self.actor_key.as_str());
-        }
-    }
-}
-
-const IDENTIFIER_FINGERPRINT_HEX_PREFIX_LEN: usize = 24;
-
-fn normalized_identifier(raw: Option<&str>) -> Option<String> {
-    raw.map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(|value| value.to_owned())
-}
-
-fn normalized_conversation_id(raw: Option<&str>) -> Option<String> {
-    normalized_identifier(raw)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct RedactedBindingScopeForLog {
-    route_session_id: String,
-    channel_id: Option<String>,
-    account_id: Option<String>,
-    conversation_id: Option<String>,
-    thread_id: Option<String>,
-}
-
-fn redact_identifier_for_log(raw: Option<&str>) -> Option<String> {
-    let normalized = normalized_identifier(raw)?;
-    Some(identifier_fingerprint(normalized.as_str()))
-}
-
-fn identifier_fingerprint(raw: &str) -> String {
-    let digest = Sha256::digest(raw.as_bytes());
-    let hex = hex::encode(digest);
-    let prefix = &hex[..IDENTIFIER_FINGERPRINT_HEX_PREFIX_LEN];
-    format!("sha256:{prefix}")
-}
-
-fn redact_binding_scope_for_log(
-    binding: Option<&AcpSessionBindingScope>,
-) -> Option<RedactedBindingScopeForLog> {
-    let binding = binding?;
-
-    Some(RedactedBindingScopeForLog {
-        route_session_id: identifier_fingerprint(binding.route_session_id.as_str()),
-        channel_id: binding.channel_id.clone(),
-        account_id: redact_identifier_for_log(binding.account_id.as_deref()),
-        conversation_id: redact_identifier_for_log(binding.conversation_id.as_deref()),
-        thread_id: redact_identifier_for_log(binding.thread_id.as_deref()),
-    })
-}
-
-fn actor_key_for_bootstrap(bootstrap: &AcpSessionBootstrap) -> String {
-    let binding = AcpSessionBindingScope::from_bootstrap(bootstrap);
-    session_actor_key(
-        bootstrap.session_key.as_str(),
-        bootstrap.conversation_id.as_deref(),
-        binding.as_ref(),
-    )
-}
-
-fn actor_key_for_metadata(metadata: &AcpSessionMetadata) -> String {
-    session_actor_key(
-        metadata.session_key.as_str(),
-        metadata.conversation_id.as_deref(),
-        metadata.binding.as_ref(),
-    )
-}
-
-fn session_actor_key(
-    session_key: &str,
-    conversation_id: Option<&str>,
-    binding: Option<&AcpSessionBindingScope>,
-) -> String {
-    if let Some(binding) = binding {
-        return format!("route:{}", binding.route_session_id);
-    }
-    if let Some(conversation_id) = normalized_conversation_id(conversation_id) {
-        return format!("conversation:{conversation_id}");
-    }
-    format!("session:{}", session_key.trim())
-}
-
-fn normalize_error_key(error: &str) -> String {
-    let trimmed = error.trim();
-    if trimmed.is_empty() {
-        return "unknown".to_owned();
-    }
-    const MAX_ERROR_KEY_LEN: usize = 120;
-    let mut truncated = String::new();
-    for (idx, ch) in trimmed.chars().enumerate() {
-        if idx >= MAX_ERROR_KEY_LEN {
-            truncated.push_str("...");
-            break;
-        }
-        truncated.push(ch);
-    }
-    truncated
-}
-
-fn projected_status_state(
-    state: AcpSessionState,
-    active_turn: bool,
-    pending_turns: usize,
-) -> AcpSessionState {
-    if !active_turn && pending_turns == 0 {
-        return state;
-    }
-
-    if matches!(
-        state,
-        AcpSessionState::Cancelling | AcpSessionState::Error | AcpSessionState::Closed
-    ) {
-        state
-    } else {
-        AcpSessionState::Busy
-    }
-}
-
-fn bump_usize_count(counts: &mut BTreeMap<String, usize>, key: &str) {
-    let entry = counts.entry(key.to_owned()).or_insert(0);
-    *entry = entry.saturating_add(1);
-}
-
-fn now_ms() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|duration| duration.as_millis().min(u128::from(u64::MAX)) as u64)
-        .unwrap_or(0)
 }
 
 #[cfg(test)]
