+++ crates/app/src/tools/mod.rs
@@ -1,7 +1,6 @@
 #[cfg(test)]
 use std::cell::Cell;
 use std::{
-    borrow::Cow,
     collections::{BTreeMap, BTreeSet},
     ffi::OsString,
     future::Future,
@@ -15,8 +14,8 @@ use serde_json::{Value, json};
 #[cfg(test)]
 use tool_search::searchable_entry_from_provider_definition;
 use tool_search::{
-    RankedSearchableToolEntry, SearchableToolEntry, rank_searchable_entries,
-    searchable_entry_from_descriptor,
+    SearchableToolEntry, execute_tool_search_tool_with_config, searchable_entry_from_descriptor,
+    tool_search_entry_is_runtime_usable,
 };
 
 use crate::KernelContext;
@@ -64,6 +63,7 @@ mod session_search;
 mod shell;
 pub mod shell_policy_ext;
 mod shell_request_prep;
+mod tool_lease;
 mod tool_lease_authority;
 mod tool_search;
 // Browser reuses the shared SSRF and HTML helpers from web_fetch even when the
@@ -100,6 +100,14 @@ pub(crate) use shell_request_prep::{
     TOOL_SEARCH_GRANTED_CAPABILITIES_FIELD, inject_tool_lease_binding,
     normalize_shell_payload_for_request, normalize_shell_request_for_execution,
     prepare_kernel_tool_request,
+};
+pub(crate) use tool_lease::{
+    bridge_provider_tool_call_with_scope, execute_tool_invoke_tool_with_config, issue_tool_lease,
+    resolve_tool_invoke_request,
+};
+#[cfg(test)]
+pub(crate) use tool_lease::{
+    synthesize_test_provider_tool_call, synthesize_test_provider_tool_call_with_scope,
 };
 #[cfg(any(
     feature = "tool-http",
@@ -1244,32 +1252,6 @@ fn feishu_searchable_entries() -> Vec<SearchableToolEntry> {
         .collect()
 }
 
-fn search_tool_view_from_payload(
-    payload: &serde_json::Map<String, Value>,
-    config: &runtime_config::ToolRuntimeConfig,
-) -> ToolView {
-    let visible_tool_names = if trusted_internal_tool_payload_enabled() {
-        payload
-            .get(LOONGCLAW_INTERNAL_TOOL_CONTEXT_KEY)
-            .and_then(|body| body.get(LOONGCLAW_INTERNAL_TOOL_SEARCH_KEY))
-            .and_then(|body| body.get(LOONGCLAW_INTERNAL_TOOL_SEARCH_VISIBLE_TOOL_IDS_KEY))
-            .and_then(Value::as_array)
-            .map(|tool_names| {
-                tool_names
-                    .iter()
-                    .filter_map(Value::as_str)
-                    .map(canonical_tool_name)
-                    .collect::<Vec<_>>()
-            })
-    } else {
-        None
-    };
-    match visible_tool_names {
-        Some(visible_tool_names) => ToolView::from_tool_names(visible_tool_names),
-        None => full_runtime_tool_view_for_runtime_config(config),
-    }
-}
-
 fn runtime_discoverable_tool_entries(
     config: &runtime_config::ToolRuntimeConfig,
     visible_tool_view: Option<&ToolView>,
@@ -1313,393 +1295,6 @@ pub fn tool_parameter_schema_types() -> BTreeMap<String, BTreeMap<String, &'stat
     }
     tools_by_name
 }
-
-fn execute_tool_search_tool_with_config(
-    request: ToolCoreRequest,
-    config: &runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = request
-        .payload
-        .as_object()
-        .ok_or_else(|| "tool.search payload must be an object".to_owned())?;
-    let query = tool_search_query_from_payload(payload).map(Cow::into_owned);
-    let requested_exact_tool_id = payload
-        .get("exact_tool_id")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned);
-    let exact_tool_id = requested_exact_tool_id
-        .as_deref()
-        .map(canonical_tool_name)
-        .map(str::to_owned);
-
-    let limit = payload
-        .get("limit")
-        .and_then(Value::as_u64)
-        .map(|value| value.clamp(1, 8) as usize)
-        .unwrap_or(5);
-    let granted_capabilities = payload
-        .get(TOOL_SEARCH_GRANTED_CAPABILITIES_FIELD)
-        .cloned()
-        .and_then(|value| serde_json::from_value::<BTreeSet<Capability>>(value).ok());
-    let visible_tool_view = search_tool_view_from_payload(payload, config);
-
-    let searchable_entries = runtime_discoverable_tool_entries(config, Some(&visible_tool_view))
-        .into_iter()
-        .filter(|entry| {
-            tool_search_entry_is_capability_usable(
-                entry.canonical_name.as_str(),
-                granted_capabilities.as_ref(),
-            )
-        })
-        .collect::<Vec<_>>();
-    let exact_match_entry = exact_tool_id.as_ref().and_then(|exact_tool_id| {
-        searchable_entries
-            .iter()
-            .find(|entry| entry.canonical_name == *exact_tool_id)
-            .cloned()
-    });
-    let exact_match_found = exact_match_entry.is_some();
-    let mut diagnostics_reason = None;
-    let results: Vec<Value> = if let Some(entry) = exact_match_entry {
-        let why = Vec::new();
-        let entry_json = tool_search_result_entry_json(&entry, why, payload)?;
-        vec![entry_json]
-    } else if let Some(query) = query.as_deref() {
-        let ranking = rank_searchable_entries(searchable_entries, query, limit);
-        diagnostics_reason = ranking.diagnostics_reason;
-
-        ranking
-            .results
-            .into_iter()
-            .map(|ranked_entry| {
-                let RankedSearchableToolEntry { entry, why } = ranked_entry;
-
-                tool_search_result_entry_json(&entry, why, payload)
-            })
-            .collect::<Result<Vec<_>, _>>()?
-    } else {
-        let ranking = rank_searchable_entries(searchable_entries, "", limit);
-        diagnostics_reason = ranking.diagnostics_reason;
-
-        ranking
-            .results
-            .into_iter()
-            .map(|ranked_entry| {
-                let RankedSearchableToolEntry { entry, why } = ranked_entry;
-
-                tool_search_result_entry_json(&entry, why, payload)
-            })
-            .collect::<Result<Vec<_>, _>>()?
-    };
-    let diagnostics = tool_search_diagnostics_json(
-        requested_exact_tool_id.as_deref(),
-        exact_match_found,
-        query.as_deref(),
-        diagnostics_reason,
-    );
-    let response_exact_tool_id = if exact_match_found {
-        exact_tool_id
-    } else {
-        requested_exact_tool_id
-    };
-
-    Ok(ToolCoreOutcome {
-        status: "ok".to_owned(),
-        payload: json!({
-            "adapter": "core-tools",
-            "tool_name": request.tool_name,
-            "query": query,
-            "exact_tool_id": response_exact_tool_id,
-            "returned": results.len(),
-            "results": results,
-            "diagnostics": diagnostics,
-        }),
-    })
-}
-
-fn tool_search_result_entry_json(
-    entry: &SearchableToolEntry,
-    why: Vec<String>,
-    payload: &serde_json::Map<String, Value>,
-) -> Result<Value, String> {
-    let lease = tool_lease_authority::issue_tool_lease(entry.canonical_name.as_str(), payload)?;
-    let entry_json = json!({
-        "tool_id": entry.canonical_name,
-        "summary": entry.summary,
-        "search_hint": entry.search_hint,
-        "argument_hint": entry.argument_hint,
-        "required_fields": entry.required_fields,
-        "required_field_groups": entry.required_field_groups,
-        "schema_preview": entry.schema_preview,
-        "tags": entry.tags,
-        "why": why,
-        "lease": lease,
-    });
-    Ok(entry_json)
-}
-
-fn tool_search_diagnostics_json(
-    requested_exact_tool_id: Option<&str>,
-    exact_match_found: bool,
-    query: Option<&str>,
-    diagnostics_reason: Option<&str>,
-) -> Value {
-    if let Some(requested_exact_tool_id) = requested_exact_tool_id {
-        if exact_match_found {
-            return Value::Null;
-        }
-
-        return json!({
-            "reason": "exact_tool_id_not_visible",
-            "requested_tool_id": requested_exact_tool_id,
-        });
-    }
-
-    if let Some(reason) = diagnostics_reason {
-        let diagnostics_query = query.unwrap_or_default();
-
-        return json!({
-            "reason": reason,
-            "query": diagnostics_query,
-        });
-    }
-
-    Value::Null
-}
-
-fn tool_search_query_from_payload(
-    payload: &serde_json::Map<String, Value>,
-) -> Option<Cow<'_, str>> {
-    const QUERY_KEYS: &[&str] = &["query", "input", "text", "prompt", "keyword", "keywords"];
-
-    for key in QUERY_KEYS {
-        let Some(value) = payload.get(*key) else {
-            continue;
-        };
-
-        if let Some(query) = tool_search_query_from_value(value) {
-            return Some(query);
-        }
-    }
-
-    None
-}
-
-fn tool_search_query_from_value(value: &Value) -> Option<Cow<'_, str>> {
-    let string_value = value.as_str();
-    if let Some(string_value) = string_value {
-        let trimmed_value = string_value.trim();
-        if !trimmed_value.is_empty() {
-            return Some(Cow::Borrowed(trimmed_value));
-        }
-    }
-
-    let values = value.as_array()?;
-    let joined_value = join_tool_search_query_values(values);
-    if joined_value.is_empty() {
-        return None;
-    }
-
-    Some(Cow::Owned(joined_value))
-}
-
-fn join_tool_search_query_values(values: &[Value]) -> String {
-    let mut query_parts = Vec::new();
-
-    for value in values {
-        let query_part = tool_search_query_part(value);
-        if query_part.is_empty() {
-            continue;
-        }
-
-        query_parts.push(query_part);
-    }
-
-    query_parts.join(" ")
-}
-
-fn tool_search_query_part(value: &Value) -> String {
-    let string_value = value.as_str();
-    if let Some(string_value) = string_value {
-        return string_value.trim().to_owned();
-    }
-
-    value.to_string()
-}
-
-fn tool_search_entry_is_runtime_usable(
-    tool_name: &str,
-    config: &runtime_config::ToolRuntimeConfig,
-) -> bool {
-    match tool_name {
-        "shell.exec" => {
-            !config.shell_allow.is_empty()
-                || matches!(
-                    config.shell_default_mode,
-                    crate::tools::shell_policy_ext::ShellPolicyDefault::Allow
-                )
-        }
-        "bash.exec" => config.bash_exec.is_discoverable(),
-        "external_skills.fetch"
-        | "external_skills.install"
-        | "external_skills.inspect"
-        | "external_skills.invoke"
-        | "external_skills.list"
-        | "external_skills.remove" => config.external_skills.enabled,
-        #[cfg(feature = "tool-file")]
-        "memory_search" => memory_tools::memory_corpus_available(config),
-        #[cfg(feature = "tool-file")]
-        "memory_get" => memory_tools::workspace_memory_corpus_available(config),
-        _ => true,
-    }
-}
-
-fn tool_search_entry_is_capability_usable(
-    tool_name: &str,
-    granted_capabilities: Option<&BTreeSet<Capability>>,
-) -> bool {
-    let Some(granted_capabilities) = granted_capabilities else {
-        return true;
-    };
-    let required = required_capabilities_for_tool_name_and_payload(tool_name, &json!({}));
-    required
-        .iter()
-        .all(|capability| granted_capabilities.contains(capability))
-}
-
-pub(crate) fn resolve_tool_invoke_request(
-    request: &ToolCoreRequest,
-) -> Result<(ResolvedToolExecution, ToolCoreRequest), String> {
-    if canonical_tool_name(request.tool_name.as_str()) != "tool.invoke" {
-        return Err(format!(
-            "tool_invoke_required: expected `tool.invoke`, got `{}`",
-            request.tool_name
-        ));
-    }
-
-    let payload = request
-        .payload
-        .as_object()
-        .ok_or_else(|| "tool.invoke payload must be an object".to_owned())?;
-    let tool_id = payload
-        .get("tool_id")
-        .and_then(Value::as_str)
-        .map(canonical_tool_name)
-        .ok_or_else(|| "tool.invoke requires payload.tool_id".to_owned())?;
-    let lease = payload
-        .get("lease")
-        .and_then(Value::as_str)
-        .ok_or_else(|| "tool.invoke requires payload.lease".to_owned())?;
-    let mut arguments = payload
-        .get("arguments")
-        .cloned()
-        .unwrap_or_else(|| json!({}));
-    {
-        let arguments_object = arguments
-            .as_object_mut()
-            .ok_or_else(|| "tool.invoke payload.arguments must be an object".to_owned())?;
-        if let Some(internal_context) = payload.get(LOONGCLAW_INTERNAL_TOOL_CONTEXT_KEY) {
-            merge_trusted_internal_tool_context_into_arguments(arguments_object, internal_context)?;
-        }
-    }
-
-    let resolved = resolve_tool_execution(tool_id)
-        .ok_or_else(|| format!("tool_not_found: unknown tool `{tool_id}`"))?;
-    let resolved_tool_name = resolved.canonical_name;
-    if is_provider_exposed_tool_name(resolved_tool_name) {
-        return Err(format!(
-            "tool_not_provider_exposed: {} must be called directly as a core tool",
-            resolved_tool_name
-        ));
-    }
-    tool_lease_authority::validate_tool_lease(resolved_tool_name, lease, payload)?;
-
-    Ok((
-        resolved,
-        ToolCoreRequest {
-            tool_name: resolved_tool_name.to_owned(),
-            payload: arguments,
-        },
-    ))
-}
-
-fn execute_tool_invoke_tool_with_config(
-    request: ToolCoreRequest,
-    config: &runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let inner_arguments = request.payload.get("arguments").unwrap_or(&Value::Null);
-    ensure_untrusted_payload_does_not_use_reserved_internal_tool_context(
-        request.tool_name.as_str(),
-        inner_arguments,
-        "payload.arguments",
-    )?;
-    let (entry, effective_request) = resolve_tool_invoke_request(&request)?;
-    match entry.execution_kind {
-        ToolExecutionKind::Core => {
-            execute_discoverable_tool_core_with_config(effective_request, config)
-        }
-        ToolExecutionKind::App => Err(format!(
-            "tool_requires_app_dispatcher: {}",
-            entry.canonical_name
-        )),
-    }
-}
-
-#[allow(dead_code)]
-pub(crate) fn bridge_provider_tool_call_with_scope(
-    tool_name: &str,
-    args_json: Value,
-    session_id: Option<&str>,
-    turn_id: Option<&str>,
-) -> (String, Value) {
-    let canonical_name = canonical_tool_name(tool_name).to_owned();
-    let Some(entry) = catalog::find_tool_catalog_entry(canonical_name.as_str()) else {
-        return (canonical_name, args_json);
-    };
-    if !entry.is_discoverable() {
-        return (canonical_name, args_json);
-    }
-    let mut lease_payload = serde_json::Map::new();
-    inject_tool_lease_binding(&mut lease_payload, None, session_id, turn_id);
-    let lease_result = tool_lease_authority::issue_tool_lease(entry.canonical_name, &lease_payload);
-    let lease = match lease_result {
-        Ok(lease) => lease,
-        Err(error) => {
-            let invalid_lease = format!("tool-lease-error:{error}");
-            invalid_lease
-        }
-    };
-    let mut outer_payload = serde_json::Map::new();
-    outer_payload.insert("tool_id".to_owned(), json!(entry.canonical_name));
-    outer_payload.insert("lease".to_owned(), json!(lease));
-    outer_payload.insert("arguments".to_owned(), args_json);
-    for (key, value) in lease_payload {
-        outer_payload.insert(key, value);
-    }
-    ("tool.invoke".to_owned(), Value::Object(outer_payload))
-}
-
-#[cfg(test)]
-#[allow(dead_code)]
-pub(crate) fn synthesize_test_provider_tool_call(
-    tool_name: &str,
-    args_json: Value,
-) -> (String, Value) {
-    bridge_provider_tool_call_with_scope(tool_name, args_json, None, None)
-}
-
-#[cfg(test)]
-pub(crate) fn synthesize_test_provider_tool_call_with_scope(
-    tool_name: &str,
-    args_json: Value,
-    session_id: Option<&str>,
-    turn_id: Option<&str>,
-) -> (String, Value) {
-    bridge_provider_tool_call_with_scope(tool_name, args_json, session_id, turn_id)
-}
-
 fn tool_function_name(tool: &Value) -> &str {
     tool.get("function")
         .and_then(|value| value.get("name"))
